// projects/ch-web/src/app/models/center-for-week.ts
import { Session } from "./session";

export interface CenterForWeek {
    center_id: number;
    name: string;
    address: string;
    state_name: string;
    district_name: string;
    block_name: string;
    pincode: number;
    lat: number;
    long: number;
    from: string;
    to: string;
    fee_type: string;
    sessions: Session[]
}